Exploding Fish Brick
Implemented the coding for the exploding fish brick, which successfully explodes the bricks around it with the special tag. Still having an issue with adding too many points, though. Seems to be adding the points of every object in the array (even if it has been destroyed already) rather than just the ones destroyed by the fish.